requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
django==3.2
gunicorn==20.0.4
django-heroku==0.3.1
torch==1.11.0+cpu
djangorestframework
scikit-learn
tensorflow-cpu==2.8.0